matrix: move compute_full_hessian
