+++ opt_adapt/matrix.py
@@ -1,8 +1,9 @@
 import firedrake as fd
+import firedrake_adjoint as fd_adj
 import numpy as np
 
 
-__class__ = ["Matrix"]
+__class__ = ["Matrix", "compute_full_hessian"]
 
 
 class Matrix:
@@ -111,3 +112,27 @@ class Matrix:
         solution = fd.Function(self.function_space)
         solution.dat.data[:] = u.reshape(f.shape)
         return solution
+
+
+def compute_full_hessian(J, u):
+    """
+    Compute the full Hessian of a functional
+    w.r.t. a control.
+
+    :arg J: the functional
+    :arg u: the :class:`Control`
+    """
+    if not isinstance(u, fd_adj.Control):
+        raise ValueError(f"Second argument should be a Control, not {type(u)}")
+    Jhat = fd_adj.ReducedFunctional(J, u)
+    fs = u.data().function_space()
+    H = Matrix(fs)
+    Rspace = fs.ufl_element().family() == "Real"
+    if u.block_variable.adj_value is None:
+        Jhat.derivative()
+    if Rspace:
+        h = fd.Function(fs).assign(1.0)
+        H.set(Jhat.hessian(h).dat.data)
+    else:
+        raise NotImplementedError("Full Hessian only supported for R-space")
+    return H
